+++ built_value/test/regexp_serializer_test.dart
@@ -1,0 +1,42 @@
+// Copyright (c) 2015, Google Inc. Please see the AUTHORS file for details.
+// All rights reserved. Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+import 'package:built_value/serializer.dart';
+import 'package:test/test.dart';
+
+void main() {
+  final serializers = new Serializers();
+
+  group('RegExp with known specifiedType', () {
+    final data = RegExp(r'\d+, [A-Z](foo)?');
+    final serialized = r'\d+, [A-Z](foo)?';
+    final specifiedType = const FullType(RegExp);
+
+    test('can be serialized', () {
+      expect(serializers.serialize(data, specifiedType: specifiedType),
+          serialized);
+    });
+
+    test('can be deserialized', () {
+      expect(serializers.deserialize(serialized, specifiedType: specifiedType),
+          data);
+    });
+  });
+
+  group('String with unknown specifiedType', () {
+    final data = RegExp('testing, testing');
+    final serialized = ['RegExp', 'testing, testing'];
+    final specifiedType = FullType.unspecified;
+
+    test('can be serialized', () {
+      expect(serializers.serialize(data, specifiedType: specifiedType),
+          serialized);
+    });
+
+    test('can be deserialized', () {
+      expect(serializers.deserialize(serialized, specifiedType: specifiedType),
+          data);
+    });
+  });
+}