+++ end_to_end_test/lib/values.dart
@@ -276,6 +276,7 @@ abstract class PrimitivesValue
   String get string;
   DateTime get dateTime;
   Duration get duration;
+  RegExp get regExp;
   Uri get uri;
   BigInt get bigInt;