+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Make it possible to merge `Serializers` instances: add a `serializers`
   getter that returns the installed serializers.
 - Allow `analyzer` 0.34.0.
+- Add serializer for `RegExp` fields.
 
 # 6.2.0